+++ src/IO/sion-fti.c
@@ -178,7 +178,7 @@ int FTI_WriteSionData(FTIT_dataset *data, void *fd) {
 }
 
 
-size_t FTI_GetSionFilePos(void *fileDesc) {
+int64_t FTI_GetSionFilePos(void *fileDesc) {
     WriteSionInfo_t *fd  = (WriteSionInfo_t *) fileDesc;
     return fd->loffset;
 }
